rollback useless modifications
