+++ src/main/java/com/powsybl/openloadflow/knitro/solver/AbstractKnitroProblem.java
@@ -23,6 +23,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
+import java.util.stream.IntStream;
 
 /**
  * Abstract base class for Knitro problem definitions, providing common functionality.
@@ -210,7 +211,7 @@ public abstract class AbstractKnitroProblem extends KNProblem {
             try {
                 if (knitroParameters.getGradientUserRoutine() == 1) {
                     // Dense method: all non-linear constraints are considered as a function of all variables.
-                    NonLinearExternalSolverUtils.buildDenseJacobianMatrix(numberOfPowerFlowVariables, listNonLinearConsts,
+                    buildDenseJacobianMatrix(numberOfPowerFlowVariables, listNonLinearConsts,
                             listNonZerosCtsDense, listNonZerosVarsDense);
                     this.setJacNnzPattern(listNonZerosCtsDense, listNonZerosVarsDense);
                 } else if (knitroParameters.getGradientUserRoutine() == 2) {
@@ -223,12 +224,38 @@ public abstract class AbstractKnitroProblem extends KNProblem {
                 throw new PowsyblException("Failed to set Jacobian pattern in Knitro problem", e);
             }
             // Set the callback for gradient evaluations if the user directly passes the Jacobian to the solver.
-            this.setGradEvalCallback(createGradientCallback(
-                    jacobianMatrix,
-                    listNonZerosCtsDense,
-                    listNonZerosVarsDense,
-                    listNonZerosCtsSparse,
-                    listNonZerosVarsSparse));
+            this.setGradEvalCallback(createGradientCallback(jacobianMatrix, listNonZerosCtsDense, listNonZerosVarsDense,
+                    listNonZerosCtsSparse, listNonZerosVarsSparse));
+        }
+    }
+
+
+    /**
+     * Builds the row and column index lists corresponding to the dense Jacobian structure,
+     * assuming each non-linear constraint is derived with respect to every variable.
+     *
+     * @param numVars               Total number of variables.
+     * @param listNonLinearConsts   List of non-linear constraint indices.
+     * @param listNonZerosCtsDense  Output list to receive constraint indices for non-zero Jacobian entries.
+     * @param listNonZerosVarsDense Output list to receive variable indices for non-zero Jacobian entries.
+     */
+    public void buildDenseJacobianMatrix(
+            int numVars,
+            List<Integer> listNonLinearConsts,
+            List<Integer> listNonZerosCtsDense,
+            List<Integer> listNonZerosVarsDense) {
+
+        // Each non-linear constraint will have a partial derivative with respect to every variable
+        for (Integer constraintId : listNonLinearConsts) {
+            for (int varIndex = 0; varIndex < numVars; varIndex++) {
+                listNonZerosCtsDense.add(constraintId);
+            }
+        }
+
+        // We repeat the full list of variables for each non-linear constraint
+        List<Integer> variableIndices = IntStream.range(0, numVars).boxed().toList();
+        for (int i = 0; i < listNonLinearConsts.size(); i++) {
+            listNonZerosVarsDense.addAll(variableIndices);
         }
     }
