+++ src/main/java/com/powsybl/openloadflow/knitro/solver/KnitroStatus.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2025, Artelys (http://www.artelys.com/)
+ * Copyright (c) 2024, Artelys (http://www.artelys.com/)
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
